[#1509] Resolve implicit numeric conversion warnings
Make time and integer-to-floating conversions explicit, use dedicated
nanosecond conversion helpers, and preserve integer storage calculations.
Exclude the broader signedness cleanup for separate review.

+++ src/fswAlgorithms/sensorInterfaces/simpleInstrumentController/simpleInstrumentController.c
@@ -155,7 +155,7 @@ Update_simpleInstrumentController(simpleInstrumentControllerConfig* configData, 
                                 "simpleInstrumentController: allowedTime is negative and has been set to zero.");
                     }
 
-                    double elapsedTime = callTime - configData->constraintStartTime;
+                    double elapsedTime = (double) callTime - configData->constraintStartTime;
 
                     // Determine the effective time to image: cannot exceed allowedTime
                     double effectiveImageTime = (configData->acquisitionTime > configData->allowedTime)
